Change value of return to float

+++ python_async_comprehension/2-measure_runtime.py
@@ -5,7 +5,7 @@ import time
 async_comprehension = __import__('1-async_comprehension').async_comprehension
 
 
-async def measure_runtime() -> int:
+async def measure_runtime() -> float:
     """ Execute async_comprehension four times in parallel
     using asyncio.gather"""
     start = time.time()
